% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRSquared.R
\name{MRSq}
\alias{MRSq}
\title{Performs multiply robust Mendelian randomization analysis with many invalid instruments}
\usage{
MRSq(Y, A, G.mat, k = 5)
}
\arguments{
\item{Y}{length n continuous outcome vector}

\item{A}{length n continuous exposure vector}

\item{G.mat}{n by K binary SNP matrix}

\item{k}{2<=k<=K, the number of valid IVs}
}
\value{
Point estimate of causal effect of A on Y and standard error
}
\description{
Performs multiply robust Mendelian randomization analysis with many invalid instruments
}
